make the Prolog path planner the standard path planner
