+++ src/prolog/__init__.py
@@ -1,22 +1,22 @@
 """
-Global Optimized Prolog Path Planner package for Taiat.
+Taiat Path Planner package.
 
-This package provides global optimized Prolog-based path planning functionality for determining
+This package provides Prolog-based path planning functionality for determining
 optimal execution sequences in Taiat queries.
 """
 
-from .optimized_prolog_interface import (
-    OptimizedPrologPathPlanner,
-    plan_taiat_path_optimized,
-    plan_taiat_path_global_optimized,
-    get_global_optimized_planner,
+from .taiat_path_planner import (
+    TaiatPathPlanner,
+    plan_taiat_path,
+    plan_taiat_path_global,
+    get_global_planner,
 )
 
 __all__ = [
-    "OptimizedPrologPathPlanner",
-    "plan_taiat_path_optimized",
-    "plan_taiat_path_global_optimized",
-    "get_global_optimized_planner",
+    "TaiatPathPlanner",
+    "plan_taiat_path",
+    "plan_taiat_path_global",
+    "get_global_planner",
 ]
 
 __version__ = "2.0.0"
